reorganizing commands and implementing public urls on API
