--- src/ItsyBitseList.Core/WishlistAggregate/Wishlists/Commands/DeleteItemInWishlist/DeleteItemInWishlistHandler.cs
@@ -1,32 +1,0 @@
-﻿using ItsyBitseList.Core.Interfaces.Persistence;
-using ItsyBitseList.Core.WishlistCollectionAggregate;
-using MediatR;
-
-namespace ItsyBitseList.Core.WishlistAggregate.Wishlists.Commands.DeleteItemInWishlist
-{
-    public record DeleteItemInWishlistCommand(
-        Guid WishlistId,
-        Guid ItemId
-    ) : IRequest;
-
-    public class DeleteItemInWishlistHandler : IRequestHandler<DeleteItemInWishlistCommand>
-    {
-        public IAsyncRepository<Wishlist> _repository;
-        public DeleteItemInWishlistHandler(IAsyncRepository<Wishlist> repository)
-        {
-            this._repository = repository;
-        }
-        public async Task Handle(DeleteItemInWishlistCommand request, CancellationToken cancellationToken)
-        {
-            var wishlist = await _repository.GetByIdAsync(request.WishlistId);
-            var itemToDelete = wishlist.Items.First(item => item.Id == request.ItemId);
-            if (itemToDelete is null || itemToDelete.WishlistId != request.WishlistId)
-            {
-                throw new UnauthorizedAccessException("Item not found in wishlist");
-            }
-            wishlist.Remove(request.ItemId);
-            await _repository.UpdateAsync(wishlist);
-        }
-
-    }
-}
