+++ tests/ItsyBitseList.IntegrationTests/AzureTableRepository/PublicWishlistEnpointTests.cs
@@ -1,7 +1,7 @@
 ﻿using FluentAssertions;
-using ItsyBitseList.Core.WishlistAggregate.Wishlists.Queries.GetItemInWishlist;
 using ItsyBitseList.Core.WishlistCollectionAggregate;
 using Microsoft.AspNetCore.Mvc.Testing;
+using static ItsyBitseList.Core.WishlistAggregate.Wishlists.Queries.GetItemInWishlist.GetItemInWishlist;
 
 namespace ItsyBitseList.IntegrationTests.AzureTableRepository
 {
@@ -12,11 +12,12 @@ namespace ItsyBitseList.IntegrationTests.AzureTableRepository
         [Fact]
         public async Task CanPromiseItem()
         {
-            (var location, var wishlist) = await CreateAndAddItems("My Wishlist Item");
+            (var location, var wishlist, var item) = await CreateAndAddItems("My Wishlist Item");
+            location = $"public/{wishlist.PublicId}/item/{item.Id}";
             HttpResponseMessage response = await PromiseItem(location);
             var itemResponse = await _client.GetAsync(location);
-            var item = await Parse<ItemDetails>(itemResponse);
-
+            item = await Parse<ItemDetails>(itemResponse);
+            
             response.EnsureSuccessStatusCode();
             item.State.Should().Be(State.Promised);
 
@@ -26,14 +27,15 @@ namespace ItsyBitseList.IntegrationTests.AzureTableRepository
         [Fact]
         public async Task CanRevertPromise()
         {
-            (var location, _) = await CreateAndAddItems("My Wishlist Item");
+            (var location, var wishlist, var item) = await CreateAndAddItems("My Wishlist Item");
+            location = $"public/{wishlist.PublicId}/item/{item.Id}";
             HttpResponseMessage response = await PromiseItem(location);
 
             var promiseKey = await response.Content.ReadAsStringAsync();
             response = await RevertPromiseItem(location, promiseKey);
 
             var itemResponse = await _client.GetAsync(location);
-            var item = await Parse<ItemDetails>(itemResponse);
+            item = await Parse<ItemDetails>(itemResponse);
 
             response.EnsureSuccessStatusCode();
             item.State.Should().Be(State.Wished);
