+++ tests/ItsyBitseList.IntegrationTests/readme.md
@@ -1,0 +1,4 @@
+﻿# Run Azurite in to run integration tests
+```
+azurite --silent --location c:\azurite --debug c:\azurite\debug.log
+```
